Modified ST.h and ST.c. Old Coord structure contained 2 short int fields, while coordinates in the binary file are written as integers. Also STsearch and STinsert have been udapted to the new data structure.

+++ src/ST.c
@@ -63,7 +63,7 @@ int STmaxSize(ST st) {
   return n;
 }
 
-void STinsert(ST st, short int coord1, short int coord2, int i) {
+void  STinsert(ST st, int coord1, int coord2, int i){
   pthread_mutex_lock(st->me);
   if (i >= st->maxN) {
     st->coord = realloc(st->coord, (2*st->maxN)*sizeof(Coord));
@@ -74,13 +74,14 @@ void STinsert(ST st, short int coord1, short int coord2, int i) {
 
   Coord c;
   c = malloc(sizeof(*c));
-  c->c1 = coord1; c->c2 = coord2;
+  c->c1 = coord1; 
+  c->c2 = coord2;
   st->coord[i] = c;
   st->N++;
   pthread_mutex_unlock(st->me);
 }
 
-int STsearch(ST st, short int coord1, short int coord2) {
+int STsearch(ST st, int coord1, int coord2) {
   int i;
   for(i = 0; i  < st->N; i++)
     if( st->coord[i] != NULL
